+++ src/pages/incidents/incidents.ts
@@ -1,13 +1,15 @@
 import { Options, Vue } from 'vue-class-component'
 
 import { IncidentsService, StreamService, VuexService } from '@/services'
-import { Incident, Project, Stream } from '@/types'
+import { Incident, Project, Stream, Pagination } from '@/types'
 import { formatDayWithoutTime, formatDiffFromNow } from '@/utils'
 import IncidentsTableRows from '../../components/incidents-table/incidents-table.vue'
+import PaginationComponent from '../../components/pagination/pagination.vue'
 
 @Options({
   components: {
-    IncidentsTableRows
+    IncidentsTableRows,
+    PaginationComponent
   }
 })
 export default class IncidentsPage extends Vue {
@@ -18,23 +20,29 @@ export default class IncidentsPage extends Vue {
   public isLoading = false
   public incidents: Incident[] = []
   public streamsData: Stream[] = []
-  public originalData: Incident[] = []
+
+  public isOpenedIncidents: string | string[] | undefined
   public limit = 2
   public alertsLabel = ''
+  public paginationSettings: Pagination = {
+    total: 0,
+    limit: 10,
+    offset: 0,
+    page: 1
+  }
 
   updated (): void {
     if (this.selectedProject !== undefined && this.selectedProject.id !== this.$route.params.projectId) {
       this.getSelectedProject()
       this.isLoading = true
       void this.getIncidentsData(this.$route.params.projectId)
-    } else if (this.$route.params.isOpenedIncidents !== undefined) {
-      this.incidents = this.originalData.filter(incident => {
-        if (this.$route.params.isOpenedIncidents === 'false') {
-          return incident.closedAt
-        } else return !incident.closedAt
-      })
-    } else {
-      this.incidents = this.originalData
+    }
+    const temp = this.isOpenedIncidents
+    if (this.$route.params.isOpenedIncidents !== undefined && temp !== this.$route.params.isOpenedIncidents) {
+      this.isOpenedIncidents = this.$route.params.isOpenedIncidents
+      this.isLoading = true
+      this.resetPaginationData()
+      void this.getIncidentsData(this.$route.params.projectId, this.$route.params.isOpenedIncidents === 'false')
     }
   }
 
@@ -44,6 +52,15 @@ export default class IncidentsPage extends Vue {
     const params: string = this.$route.params.projectId as string
     void this.getStreamsData(params)
     void this.getIncidentsData(params)
+  }
+
+  public resetPaginationData (): void {
+    this.paginationSettings = {
+      total: 0,
+      limit: 10,
+      offset: 0,
+      page: 1
+    }
   }
 
   public getSelectedProject (): void {
@@ -108,14 +125,26 @@ export default class IncidentsPage extends Vue {
     await VuexService.Projects.streams.set(this.streamsData)
   }
 
-  public async getIncidentsData (projectId: string | string[]): Promise<void> {
-    this.originalData = await IncidentsService.getIncidents({ projects: projectId })
-    this.incidents = this.formatIncidents()
+  public async getPage (): Promise<void> {
+    if (this.$route.params.isOpenedIncidents !== undefined) {
+      await this.getIncidentsData(this.$route.params.projectId, this.$route.params.isOpenedIncidents === 'false')
+    } else await this.getIncidentsData(this.$route.params.projectId)
+  }
+
+  public async getIncidentsData (projectId: string | string[], closed?: boolean): Promise<void> {
+    const data = await IncidentsService.getIncidents({
+      projects: projectId,
+      limit: this.paginationSettings.limit,
+      offset: this.paginationSettings.offset * this.paginationSettings.limit,
+      ...closed !== undefined && { closed: closed }
+    })
+    this.paginationSettings.total = data.headers['total-items']
+    this.incidents = this.formatIncidents(data.data)
     this.isLoading = false
   }
 
-  public formatIncidents (): Incident[] {
-    return this.originalData.map((incident) => {
+  public formatIncidents (incidents: Incident[]): Incident[] {
+    return incidents.map((incident) => {
       IncidentsService.combineIncidentItems(incident)
       return incident
     })