+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.1 (2021-??-??)
 
+* **feature:** Pagination is added to the Incidents page ([CE-1499](https://jira.rfcx.org/browse/CE-1499))
 * **feature:** Homepage updates ([CE-1477](https://jira.rfcx.org/browse/CE-1477))
 * **setup:** Setup search and scroll functionality to the Projects list ([CE-1474](https://jira.rfcx.org/browse/CE-1474))
 * **feature:** User is able to download Response assets ([CE-1489](https://jira.rfcx.org/browse/CE-1489))